add proof of interchange inverse-transpose
Dear Professor Gao,

It would be helpful to show the proof of : $(A^{-1})^T = (A^T)^{-1}$

To prove this, we need to show that $A^T(A^{-1})^T = I = A^T(A^T)^{-1}$

So, using the property of $(AB)^T = B^TA^T$ 

we show: $A^T(A^{-1})^T = (A^{-1}A)^T = I^T = I$

Also: $(A^{-1})^TA^T = (AA^{-1})^T = I^T = I$

So it is interchangeable. This comment is typed in Latex, please kindly verify it in Latex.
I haven't changed your original file yet, except adding a line at 36 with a comment, and typed the proof here as a message, because students like me, who lack a solid foundation in linear algebra, often find it hard to understand the proof of symmetry
of projection matrix P, when inverse and transpose of X are interchangeable. I really appreciate your project for it helped me a lot when I am teaching myself Hayashi's Econometrics, and thank you very much!